+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-03 21:33:15
-// Commit: feat(core): implement shimmer loading skeleton in core
+// Updated at: 2026-06-05 12:40:50
+// Commit: feat(core): implement responsive core view component
 
 class coreComponentState {
-  final String commitRef = '1732';
-  final String lastMessage = 'feat(core): implement shimmer loading skeleton in core';
-  final DateTime timestamp = DateTime.parse('2026-06-03T21:33:15+05:00');
+  final String commitRef = '1759';
+  final String lastMessage = 'feat(core): implement responsive core view component';
+  final DateTime timestamp = DateTime.parse('2026-06-05T12:40:50+05:00');
 }